Add files via upload
The code is fully working now. The program runs for 13 iterations currently(which is sufficient to see the pattern). You can change rules during the iteration however you want. Currently, for every even iteration the rule is multiplied by 2 and for every odd iteration the code is divided by two, this creates a beautiful pattern. ENJOY.